Feat: New Weapon System
* Projectiule 기반으로 Weapon 시스템 다시 제작
* Handgun: 평범하게 총을 쏨
* ShiledGenerator: 회전하는 쉴드를 만듦

+++ Assets/Scripts/Map/DamagedBlock.cs
@@ -9,7 +9,7 @@ using UnityEngine.UIElements;
 public class DamagedBlock
 {
     private GameObject _gameObject;
-    private int _hp;
+    private float _hp;
     
     public bool isBroken = false;
 
@@ -66,7 +66,7 @@ public class DamagedBlock
 
     public MeshFilter GetMeshFilter() { return _meshFilter; }
 
-    public void DecreaseHP(int damage)
+    public void DecreaseHP(float damage)
     {
         _gameObject.SetActive(true);
         _shakingSequence.Restart();
